Updates input_crypto and get_ticker_data funcionts
